PLANET-4811 Add scroll padding
* Prevents the navbar covering the element we're scrolling to. We
probably want to add this site-wide, but probably best handled
distinctly so we can check if this works everywhere.
* For now this padding is tuned for non-logged in users, as the admin
area needs some additional padding which is more tricky to detect. The
scroll-padding-top needs to be on html but on that element we can't know
if we're logged in or not.